+++ tf-modules/archive/variables.tf
@@ -305,7 +305,3 @@ variable "non_complete_execution_payload_timeout" {
   type    = number
   default = 30
 }
-
-variable "region" {
-  type = string
-}