+++ tf-modules/cumulus/archive.tf
@@ -2,7 +2,6 @@ module "archive" {
   source = "../archive"
 
   prefix = var.prefix
-  region = var.region
 
   permissions_boundary_arn = var.permissions_boundary_arn